# api/openapi.yaml
openapi: 3.0.3
info:
  title: Contacts API
  version: 1.0.0
  description: >-
    Self-hosted contact management API. Supports Person records and optional
    Authentik OIDC SSO. When SSO is configured, Person records are account-scoped.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
paths:
  /auth/login:
    get:
      summary: Start Authentik login
      operationId: startLogin
      responses:
        '302': { description: Redirect to Authentik }
  /auth/callback:
    get:
      summary: Complete Authentik login
      operationId: completeLogin
      responses:
        '302': { description: Redirect to the application }
  /auth/logout:
    post:
      summary: End the current session
      operationId: logout
      responses:
        '204': { description: Session ended }
  /persons:
    get:
      summary: List persons
      operationId: listPersons
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [display_name, first_name, last_name, created_at, updated_at]
            default: display_name
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: first_name
          in: query
          schema: { type: string }
        - name: last_name
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A page of persons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonListResponse'
    post:
      summary: Create a person
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /persons/deleted:
    get:
      summary: List deleted persons
      operationId: listDeletedPersons
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [display_name, first_name, last_name, created_at, updated_at]
            default: display_name
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: A page of deleted persons
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PersonListResponse' }
  /persons/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a person by ID
      operationId: getPerson
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a person
      operationId: updatePerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      summary: Soft-delete a person (recoverable for 30 days)
      operationId: deletePerson
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
  /persons/{id}/restore:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Restore a deleted person
      operationId: restorePerson
      responses:
        '204': { description: Restored }
        '404': { $ref: '#/components/responses/NotFound' }
  /persons/{id}/permanent:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Permanently delete a deleted person
      operationId: permanentlyDeletePerson
      responses:
        '204': { description: Permanently deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /sync-accounts:
    get:
      summary: List sync accounts
      operationId: listSyncAccounts
      responses:
        '200':
          description: A list of sync accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncAccountListResponse'
    post:
      summary: Create a sync account
      operationId: createSyncAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSyncAccountInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncAccount'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /sync-accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a sync account by ID
      operationId: getSyncAccount
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncAccount'
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update a sync account
      operationId: updateSyncAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSyncAccountInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncAccount'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      summary: Delete a sync account
      operationId: deleteSyncAccount
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    contactsSession:
      type: apiKey
      in: cookie
      name: contacts_session
  schemas:
    CustomFieldValue:
      description: A scalar custom field value. Dates are ISO-8601 strings.
      oneOf:
        - type: string
        - type: number
        - type: boolean
    CustomFields:
      type: object
      maxProperties: 64
      additionalProperties:
        $ref: '#/components/schemas/CustomFieldValue'
      description: >-
        Map of snake_case keys to scalar values. Max 64 fields; key max 64
        chars; string value max 1024 chars.
    Person:
      type: object
      required: [id, first_name, last_name, display_name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        first_name: { type: string }
        middle_names:
          type: array
          items: { type: string }
        last_name: { type: string }
        display_name: { type: string }
        nickname: { type: string, nullable: true }
        pronouns: { type: string, nullable: true }
        birthdate: { type: string, format: date, nullable: true }
        phone_numbers:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 50 }
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at:
          type: string
          format: date-time
          nullable: true
    CreatePersonInput:
      type: object
      required: [first_name, last_name]
      properties:
        first_name: { type: string, maxLength: 255 }
        middle_names:
          type: array
          items: { type: string }
        last_name: { type: string, maxLength: 255 }
        nickname: { type: string, maxLength: 255 }
        pronouns: { type: string, maxLength: 255 }
        birthdate: { type: string, format: date }
        phone_numbers:
          type: array
          items: { type: string, maxLength: 50 }
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
    UpdatePersonInput:
      type: object
      properties:
        first_name: { type: string, maxLength: 255 }
        middle_names:
          type: array
          items: { type: string }
        last_name: { type: string, maxLength: 255 }
        nickname: { type: string, maxLength: 255 }
        pronouns: { type: string, maxLength: 255 }
        birthdate: { type: string, format: date }
        phone_numbers:
          type: array
          items: { type: string, maxLength: 50 }
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
    PersonListResponse:
      type: object
      required: [data, page, page_size, total, total_pages]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Person' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    SyncAccount:
      type: object
      required: [id, provider, provider_account_id, sync_frequency_minutes, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid, nullable: true }
        provider: { type: string }
        provider_account_id: { type: string }
        access_token: { type: string, nullable: true }
        refresh_token: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        scope: { type: string }
        sync_cursor: { type: string }
        sync_frequency_minutes: { type: integer }
        status: { type: string }
        last_synced_at: { type: string, format: date-time, nullable: true }
        last_error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SyncAccountListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/SyncAccount' }
    CreateSyncAccountInput:
      type: object
      required: [provider, provider_account_id]
      properties:
        provider: { type: string, maxLength: 255 }
        provider_account_id: { type: string, maxLength: 255 }
        access_token: { type: string, nullable: true }
        refresh_token: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        scope: { type: string }
        sync_cursor: { type: string }
        sync_frequency_minutes: { type: integer, minimum: 1, default: 60 }
        status: { type: string }
    UpdateSyncAccountInput:
      type: object
      properties:
        provider: { type: string, nullable: true }
        provider_account_id: { type: string, nullable: true }
        access_token: { type: string, nullable: true }
        refresh_token: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        scope: { type: string, nullable: true }
        sync_cursor: { type: string, nullable: true }
        sync_frequency_minutes: { type: integer, minimum: 1, nullable: true }
        status: { type: string, nullable: true }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }